reformat folders, connect backend to frontend using axios and cors. Also setup some template get and set controllers
