+++ package.json
@@ -1,21 +1,27 @@
 {
-  "name": "backend",
+  "name": "server",
   "version": "1.0.0",
   "description": "React web application for teaching kids how to use AI",
   "main": "server.js",
   "scripts": {
-    "start": "node backend/server.js",
-    "server": "nodemon backend/server.js"
+    "start": "node server/server.js",
+    "server": "nodemon server/server.js",
+    "client": "npm start --prefix client",
+    "dev": "concurrently \"npm run server\" \"npm run client"
   },
   "author": "Max Verhoef",
   "license": "Apache-2.0",
   "dependencies": {
+    "axios": "^1.6.8",
     "colors": "^1.4.0",
+    "concurrently": "^8.2.2",
+    "cors": "^2.8.5",
     "dotenv": "^16.4.5",
     "express": "^4.19.2",
     "express-async-handler": "^1.2.0",
     "mongodb": "^6.5.0",
     "mongoose": "^8.3.2",
-    "nodemon": "^3.1.0"
+    "nodemon": "^3.1.0",
+    "react-toastify": "^10.0.5"
   }
 }
