+++ server/server.js
@@ -2,6 +2,7 @@ const path = require('path')
 const express = require('express')
 const colors = require('colors')
 const dotenv = require('dotenv').config()
+const cors = require('cors')
 const { errorHandler } = require('./middleware/errorMiddleware')
 const connectDB = require('./config/db')
 const port = process.env.PORT || 5000
@@ -9,9 +10,10 @@ const port = process.env.PORT || 5000
 connectDB()
 
 const app = express()
-
+app.use(cors())
 app.use(express.json())
 app.use(express.urlencoded({ extended: false }))
+
 
 //app.use('/api/goals', require('./routes/goalRoutes'))
 app.use('/api/quizzes', require('./routes/quizRoutes'))
